+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unknown]
+
+### Changed
+- Bundle results are now retrieved via a callback.
+
 ## [3.0.0-alpha.3] - 2018-11-21
 
 ### Changed